refactor(status-badge): migrate shared consumers to ui/status-badge canonical API
Replace import from @/components/shared/status-badge with @/components/ui/status-badge
across 6 consumer files. Convert children pattern to label prop. Map secondary intent
to neutral (closest semantic equivalent in ui/ API).
