+++ src/components/camporees/camporee-payments-panel.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { CheckCircle2, DollarSign, Loader2, Pencil, XCircle } from "lucide-react";
 import { toast } from "sonner";
 import { Badge } from "@/components/ui/badge";
-import { StatusBadge } from "@/components/shared/status-badge";
+import { StatusBadge } from "@/components/ui/status-badge";
 import { Button } from "@/components/ui/button";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
 import {
@@ -54,21 +54,19 @@ function PaymentStatusBadge({ status }: { status?: string | null }) {
   const normalized = status.toLowerCase();
 
   if (normalized === "approved") {
-    return <StatusBadge intent="success">Aprobado</StatusBadge>;
+    return <StatusBadge intent="success" label="Aprobado" />;
   }
 
   if (normalized === "pending_approval") {
-    return <StatusBadge intent="warning">Pendiente</StatusBadge>;
+    return <StatusBadge intent="warning" label="Pendiente" />;
   }
 
   if (normalized === "rejected") {
-    return <StatusBadge intent="destructive">Rechazado</StatusBadge>;
+    return <StatusBadge intent="destructive" label="Rechazado" />;
   }
 
   return (
-    <StatusBadge intent="secondary" className="text-xs capitalize">
-      {status}
-    </StatusBadge>
+    <StatusBadge intent="neutral" label={status} className="text-xs capitalize" />
   );
 }
